# Conflicts:
#	nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/find/wageprovision/wagetable/WageTableDto.java
#	nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/find/wageprovision/wagetable/WageTableFinder.java
#	nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.dom/src/main/java/nts/uk/ctx/pr/core/dom/wageprovision/wagetable/WageTableCode.java
#	nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.dom/src/main/java/nts/uk/ctx/pr/core/dom/wageprovision/wagetable/WageTableRepository.java
#	nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.infra/src/main/java/nts/uk/ctx/pr/core/infra/entity/wageprovision/wagetable/QpbmtWageTable.java
#	nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.infra/src/main/java/nts/uk/ctx/pr/core/infra/entity/wageprovision/wagetable/QpbmtWageTablePk.java
#	nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.infra/src/main/java/nts/uk/ctx/pr/core/infra/repository/wageprovision/wagetable/JpaWageTableRepository.java

+++ nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/command/wageprovision/individualwagecontract/RegisterIndividualWageContractCommandHandler.java
@@ -23,7 +23,7 @@ public class RegisterIndividualWageContractCommandHandler extends CommandHandler
     protected void handle(CommandHandlerContext<SalIndAmountCommand> commandHandlerContext) {
         SalIndAmountCommand command = commandHandlerContext.getCommand();
         String historyId = command.getHistoryId();
-        int amountOfMoney = command.getAmountOfMoney();
+        long amountOfMoney = command.getAmountOfMoney();
         SalIndAmount salIndAmount = new SalIndAmount(historyId, amountOfMoney);
         salIndAmountRepository.update(salIndAmount);
